+++ iqm_client/CHANGELOG.rst
@@ -2,6 +2,22 @@
 Changelog
 =========
 
+Version 29.12.0 (2025-08-08)
+============================
+
+Features
+--------
+
+- Minimal changes needed to support Qiskit version 1.3 and version 1.4. :issue:`SW-944`.
+
+Version 29.11.0 (2025-08-07)
+============================
+
+Features
+--------
+
+- Added :meth:`IQMClient.get_structured_metrics` for obtaining quality metric information.
+
 Version 29.10.0 (2025-07-31)
 ============================